+++ src/interfaces/IClient.ts
@@ -1,4 +1,5 @@
 import type { Collection } from 'discord.js';
+import type DiscordOauth2 from 'discord-oauth2';
 import type { Kazagumo } from 'kazagumo';
 
 import type Button from '@/classes/Button';
@@ -11,6 +12,7 @@ export default interface IClient {
   subCommands: Collection<string, SubCommand>;
   cooldowns: Collection<string, Collection<string, number>>;
   lavalink: Kazagumo;
+  oauth: DiscordOauth2;
 
   Init(): void;
   LoadHandlers(): void;