+++ src/utils/env.ts
@@ -3,6 +3,7 @@ import { bool, cleanEnv, str } from 'envalid';
 const env = cleanEnv(process.env, {
   BOT_TOKEN: str(),
   CLIENT_ID: str(),
+  CLIENT_SECRET: str(),
   SPOTIFY_CLIENT_ID: str(),
   SPOTIFY_CLIENT_SECRET: str(),
   LAVALINK_NAMES: str(),
@@ -11,6 +12,8 @@ const env = cleanEnv(process.env, {
   LAVALINK_SECURED: str(),
   DATABASE_URL: str(),
   DISABLE_UPDATE_COMMANDS: bool({ default: false }),
+  HENRIK_VALORANT_API_KEY: str({ default: '' }),
+  BASE_URL: str({ default: 'http://localhost:3000' }),
 });
 
 export default env;