+++ CHANGELOG.md
@@ -8,6 +8,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) 
 
 ---
 
+## [**4.4.1**] - 2024-04-03
+
+### Added
+
+- Added commands: `/valorant link`
+
+### Removed
+
+- Removed **axios** and replaced it with **bun-fetch**
+
 ## [**4.4.0**] - 2024-04-02
 
 ### Changed